Introduce a superclass and pull up methods from sub-classes to resolve unutilized abstraction
