+++ lib/src/main/java/com/auth0/jwt/ExpectedClaimType.java
@@ -12,119 +12,35 @@ public interface ExpectedClaimType {
     void assertExpectedClaimType(DecodedJWT jwt, Map.Entry<String, Object> entry, Clock clock);
 }
 
-class AudienceExact implements ExpectedClaimType {
-    private void assertValidAudienceClaim(List<String> audience, List<String> values, boolean shouldContainAll) {
-        if (audience == null || (shouldContainAll && !audience.containsAll(values)) ||
-                (!shouldContainAll && Collections.disjoint(audience, values))) {
-            throw new InvalidClaimException("The Claim 'aud' value doesn't contain the required audience.");
-        }
-    }
-
+class AudienceExact extends ClaimsWithAudience {
     @Override
     public void assertExpectedClaimType(DecodedJWT jwt, Map.Entry<String, Object> entry, Clock clock) {
         assertValidAudienceClaim(jwt.getAudience(), (List<String>) entry.getValue(), true);
     }
 }
 
-class AudienceContains implements ExpectedClaimType {
-    private void assertValidAudienceClaim(List<String> audience, List<String> values, boolean shouldContainAll) {
-        if (audience == null || (shouldContainAll && !audience.containsAll(values)) ||
-                (!shouldContainAll && Collections.disjoint(audience, values))) {
-            throw new InvalidClaimException("The Claim 'aud' value doesn't contain the required audience.");
-        }
-    }
-
+class AudienceContains extends ClaimsWithAudience {
     @Override
     public void assertExpectedClaimType(DecodedJWT jwt, Map.Entry<String, Object> entry, Clock clock) {
         assertValidAudienceClaim(jwt.getAudience(), (List<String>) entry.getValue(), false);
     }
 }
 
-class PublicClaimsExpiresAt implements ExpectedClaimType {
-    private void assertValidDateClaim(Date date, long leeway, boolean shouldBeFuture, Clock clock) {
-        Date today = new Date(clock.getToday().getTime());
-        today.setTime(today.getTime() / 1000 * 1000); // truncate millis
-        if (shouldBeFuture) {
-            assertDateIsFuture(date, leeway, today);
-        } else {
-            assertDateIsPast(date, leeway, today);
-        }
-    }
-
-    private void assertDateIsFuture(Date date, long leeway, Date today) {
-        today.setTime(today.getTime() - leeway * 1000);
-        if (date != null && today.after(date)) {
-            throw new TokenExpiredException(String.format("The Token has expired on %s.", date));
-        }
-    }
-
-    private void assertDateIsPast(Date date, long leeway, Date today) {
-        today.setTime(today.getTime() + leeway * 1000);
-        if (date != null && today.before(date)) {
-            throw new InvalidClaimException(String.format("The Token can't be used before %s.", date));
-        }
-    }
-
+class PublicClaimsExpiresAt extends PublicClaimsWithDate {
     @Override
     public void assertExpectedClaimType(DecodedJWT jwt, Map.Entry<String, Object> entry, Clock clock) {
         assertValidDateClaim(jwt.getExpiresAt(), (Long) entry.getValue(), true, clock);
     }
 }
 
-class PublicClaimsIssuedAt implements ExpectedClaimType {
-    private void assertValidDateClaim(Date date, long leeway, boolean shouldBeFuture, Clock clock) {
-        Date today = new Date(clock.getToday().getTime());
-        today.setTime(today.getTime() / 1000 * 1000); // truncate millis
-        if (shouldBeFuture) {
-            assertDateIsFuture(date, leeway, today);
-        } else {
-            assertDateIsPast(date, leeway, today);
-        }
-    }
-
-    private void assertDateIsFuture(Date date, long leeway, Date today) {
-        today.setTime(today.getTime() - leeway * 1000);
-        if (date != null && today.after(date)) {
-            throw new TokenExpiredException(String.format("The Token has expired on %s.", date));
-        }
-    }
-
-    private void assertDateIsPast(Date date, long leeway, Date today) {
-        today.setTime(today.getTime() + leeway * 1000);
-        if (date != null && today.before(date)) {
-            throw new InvalidClaimException(String.format("The Token can't be used before %s.", date));
-        }
-    }
+class PublicClaimsIssuedAt extends PublicClaimsWithDate {
     @Override
     public void assertExpectedClaimType(DecodedJWT jwt, Map.Entry<String, Object> entry, Clock clock) {
         assertValidDateClaim(jwt.getIssuedAt(), (Long) entry.getValue(), false, clock);
     }
 }
 
-class PublicClaimsNotBefore implements ExpectedClaimType {
-    private void assertValidDateClaim(Date date, long leeway, boolean shouldBeFuture, Clock clock) {
-        Date today = new Date(clock.getToday().getTime());
-        today.setTime(today.getTime() / 1000 * 1000); // truncate millis
-        if (shouldBeFuture) {
-            assertDateIsFuture(date, leeway, today);
-        } else {
-            assertDateIsPast(date, leeway, today);
-        }
-    }
-
-    private void assertDateIsFuture(Date date, long leeway, Date today) {
-        today.setTime(today.getTime() - leeway * 1000);
-        if (date != null && today.after(date)) {
-            throw new TokenExpiredException(String.format("The Token has expired on %s.", date));
-        }
-    }
-
-    private void assertDateIsPast(Date date, long leeway, Date today) {
-        today.setTime(today.getTime() + leeway * 1000);
-        if (date != null && today.before(date)) {
-            throw new InvalidClaimException(String.format("The Token can't be used before %s.", date));
-        }
-    }
+class PublicClaimsNotBefore extends PublicClaimsWithDate {
     @Override
     public void assertExpectedClaimType(DecodedJWT jwt, Map.Entry<String, Object> entry, Clock clock) {
         assertValidDateClaim(jwt.getNotBefore(), (Long) entry.getValue(), false, clock);
