+++ lib/src/main/java/com/auth0/jwt/PublicClaimsWithDate.java
@@ -1,0 +1,33 @@
+package com.auth0.jwt;
+
+import com.auth0.jwt.exceptions.InvalidClaimException;
+import com.auth0.jwt.exceptions.TokenExpiredException;
+import com.auth0.jwt.interfaces.Clock;
+
+import java.util.Date;
+
+public abstract class PublicClaimsWithDate implements ExpectedClaimType {
+    private void assertDateIsFuture(Date date, long leeway, Date today) {
+        today.setTime(today.getTime() - leeway * 1000);
+        if (date != null && today.after(date)) {
+            throw new TokenExpiredException(String.format("The Token has expired on %s.", date));
+        }
+    }
+
+    private void assertDateIsPast(Date date, long leeway, Date today) {
+        today.setTime(today.getTime() + leeway * 1000);
+        if (date != null && today.before(date)) {
+            throw new InvalidClaimException(String.format("The Token can't be used before %s.", date));
+        }
+    }
+
+    protected void assertValidDateClaim(Date date, long leeway, boolean shouldBeFuture, Clock clock) {
+        Date today = new Date(clock.getToday().getTime());
+        today.setTime(today.getTime() / 1000 * 1000); // truncate millis
+        if (shouldBeFuture) {
+            assertDateIsFuture(date, leeway, today);
+        } else {
+            assertDateIsPast(date, leeway, today);
+        }
+    }
+}
